Refactor and separate code, and change UI look
